@@ -11,59 +11,62 @@
     "lint": "expo lint"
   },
   "dependencies": {
-    "@expo/vector-icons": "^14.1.0",
+    "@babel/runtime": "^7.28.4",
+    "@expo/metro-runtime": "^6.1.2",
+    "@expo/vector-icons": "^15.0.2",
     "@react-native-async-storage/async-storage": "^2.2.0",
-    "@react-native-community/datetimepicker": "8.4.1",
-    "@react-native-picker/picker": "^2.11.2",
+    "@react-native-community/datetimepicker": "8.4.4",
+    "@react-native-picker/picker": "2.11.1",
     "@react-navigation/bottom-tabs": "^7.3.10",
     "@react-navigation/elements": "^2.3.8",
     "@react-navigation/native": "^7.1.6",
     "@react-navigation/native-stack": "^7.3.10",
     "@react-navigation/stack": "^7.4.7",
     "axios": "^1.12.2",
-    "expo": "~53.0.22",
-    "expo-blur": "~14.1.5",
-    "expo-constants": "~17.1.7",
-    "expo-document-picker": "~13.1.6",
-    "expo-file-system": "~18.1.11",
-    "expo-font": "~13.3.2",
-    "expo-haptics": "~14.1.4",
-    "expo-image": "~2.4.0",
+    "babel-preset-expo": "^54.0.4",
+    "expo": "^54.0.13",
+    "expo-blur": "~15.0.7",
+    "expo-constants": "~18.0.9",
+    "expo-document-picker": "~14.0.7",
+    "expo-file-system": "~19.0.17",
+    "expo-font": "~14.0.9",
+    "expo-haptics": "~15.0.7",
+    "expo-image": "~3.0.9",
     "expo-image-picker": "^17.0.8",
-    "expo-linear-gradient": "~14.1.5",
-    "expo-linking": "~7.1.7",
+    "expo-linear-gradient": "~15.0.7",
+    "expo-linking": "~8.0.8",
     "expo-media-library": "^18.2.0",
-    "expo-router": "~5.1.5",
-    "expo-sharing": "~13.1.5",
-    "expo-speech": "~13.1.7",
-    "expo-splash-screen": "~0.30.10",
-    "expo-status-bar": "~2.2.3",
-    "expo-symbols": "~0.4.5",
-    "expo-system-ui": "~5.0.11",
-    "expo-web-browser": "~14.2.0",
+    "expo-router": "~6.0.12",
+    "expo-sharing": "~14.0.7",
+    "expo-speech": "~14.0.7",
+    "expo-splash-screen": "~31.0.10",
+    "expo-status-bar": "~3.0.8",
+    "expo-symbols": "~1.0.7",
+    "expo-system-ui": "~6.0.7",
+    "expo-web-browser": "~15.0.8",
     "i18next": "^25.5.2",
     "i18next-browser-languagedetector": "^8.2.0",
-    "react": "19.0.0",
-    "react-dom": "19.0.0",
+    "react": "19.1.0",
+    "react-dom": "19.1.0",
     "react-i18next": "^15.7.3",
-    "react-native": "0.79.5",
-    "react-native-gesture-handler": "~2.24.0",
+    "react-native": "0.81.4",
+    "react-native-gesture-handler": "~2.28.0",
     "react-native-keyboard-aware-scroll-view": "^0.9.5",
     "react-native-linear-gradient": "^2.8.3",
     "react-native-paper": "^5.14.5",
-    "react-native-reanimated": "~3.17.4",
-    "react-native-safe-area-context": "5.4.0",
-    "react-native-screens": "~4.11.1",
+    "react-native-reanimated": "~4.1.1",
+    "react-native-safe-area-context": "~5.6.0",
+    "react-native-screens": "~4.16.0",
     "react-native-vector-icons": "^10.3.0",
-    "react-native-web": "~0.20.0",
-    "react-native-webview": "13.13.5",
+    "react-native-web": "^0.21.0",
+    "react-native-webview": "13.15.0",
     "react-navigation": "^5.0.0",
     "socket.io-client": "^4.8.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.25.2",
+    "@babel/core": "^7.28.4",
     "@react-native-community/cli": "latest",
-    "@types/react": "~19.0.10",
+    "@types/react": "^19.2.2",
     "eslint": "^9.25.0",
     "eslint-config-expo": "~9.2.0",
     "typescript": "~5.8.3"
