@@ -1,5 +1,5 @@
 import React from 'react';
-import { View, ScrollView, Text, StyleSheet } from 'react-native';
+import { View, FlatList, Text, StyleSheet } from 'react-native';
 import AppointmentsSection from '@/components/ui/screen/widget/LawyerDashboard/AppointmentsSection';
 import ClientListSection from '@/components/ui/screen/widget/LawyerDashboard/ClientListSection';
 import FeedbackSection from '@/components/ui/screen/widget/LawyerDashboard/FeedbackSection';
@@ -7,18 +7,28 @@ import AnalyticsSection from '@/components/ui/screen/widget/LawyerDashboard/Anal
 import TierProgressSection from './widget/LawyerDashboard/TierProgressSection';
 
 function LawyerDashboard() {
+  const sections = [
+    { key: 'tier', component: <TierProgressSection /> },
+    { key: 'analytics', component: <AnalyticsSection /> },
+    { key: 'appointments', component: <AppointmentsSection /> },
+    { key: 'clients', component: <ClientListSection /> },
+    { key: 'feedback', component: <FeedbackSection /> },
+  ];
+
+  const renderSection = ({ item }) => (
+    <View style={{ marginBottom: 16 }}>{item.component}</View>
+  );
+
   return (
-    <ScrollView style={styles.container}>
-      <Text style={styles.title}>Lawyer Dashboard</Text>
-
-      <TierProgressSection />
-      <AnalyticsSection />
-      <AppointmentsSection />
-      <ClientListSection />
-      <FeedbackSection />
-
-      <View style={{ height: 50 }} /> 
-    </ScrollView>
+    <FlatList
+      ListHeaderComponent={<Text style={styles.title}>Lawyer Dashboard</Text>}
+      data={sections}
+      keyExtractor={(item) => item.key}
+      renderItem={renderSection}
+      contentContainerStyle={styles.container}
+      showsVerticalScrollIndicator={false}
+      ListFooterComponent={<View style={{ height: 50 }} />}
+    />
   );
 }
 
@@ -26,9 +36,8 @@ export default LawyerDashboard;
 
 const styles = StyleSheet.create({
   container: {
-    flex: 1,
-    backgroundColor: '#F9F9F9',
     padding: 16,
+    backgroundColor: '#F9F9F9',
   },
   title: {
     fontSize: 24,
