@@ -1,7 +1,7 @@
 import axios from "axios";
 
 // Change this to your backend URL (localhost or LAN IP)
-const ADMIN_API_URL = "http://localhost:3000/api/admins"; // Web backend
+const ADMIN_API_URL = "http://192.168.1.9:3000/api/admins"; // Web backend
 
 // Register a new admin
 export const registerAdmin = async (adminData) => {
