@@ -1,7 +1,7 @@
 import axios from 'axios';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 
-const API_URL = 'http://localhost:3000/api/ngo';
+const API_URL = 'http://192.168.1.9:3000/api/ngo';
 
 // Create axios instance
 const api = axios.create({
