@@ -3,12 +3,12 @@ import { Platform } from 'react-native';
 // Get API URL based on platform
 const getBaseUrl = () => {
   if (Platform.OS === 'web') {
-    return 'http://localhost:3000/api';
+    return 'http://192.168.1.9:3000/api';
   } else if (Platform.OS === 'android') {
     // Try multiple URLs for Android
-    return 'http://10.4.2.1:3000/api';
+    return 'http://192.168.1.9:3000/api';
   }
-  return 'http://localhost:3000/api';
+  return 'http://192.168.1.9:3000/api';
 };
 
 const BASE_URL = getBaseUrl();
